chore: 엔티티 패키지 구조 변경 및 각 엔티티의 레포지토리 생성
