+++ server-domain/src/main/java/com/depromeet/config/db/config/JpaAuditingConfig.java
@@ -1,0 +1,17 @@
+package com.depromeet.config.db.config;
+
+import java.util.Optional;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaAuditingConfig implements AuditorAware<String> {
+
+	@Override
+	public Optional<String> getCurrentAuditor() {
+		return Optional.of("admin");
+	}
+}
