+++ server-domain/src/main/java/com/depromeet/domains/bookmark/repository/BookmarkRepository.java
@@ -1,0 +1,7 @@
+package com.depromeet.domains.bookmark.repository;
+
+import com.depromeet.domains.bookmark.entity.Bookmark;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
+}
