+++ server-domain/src/main/java/com/depromeet/domains/follow/entity/Follow.java
@@ -1,0 +1,28 @@
+package com.depromeet.domains.follow.entity;
+
+import com.depromeet.domains.common.entity.BaseTimeEntity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Follow extends BaseTimeEntity {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long followId;
+
+	@Column(nullable = false)
+	private Long followerId;
+
+	@Column(nullable = false)
+	private Long followingId;
+}
