+++ server-domain/src/main/java/com/depromeet/domains/follow/repository/FollowRepository.java
@@ -1,0 +1,7 @@
+package com.depromeet.domains.follow.repository;
+
+import com.depromeet.domains.follow.entity.Follow;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FollowRepository extends JpaRepository<Follow, Long> {
+}
