+++ server-domain/src/main/java/com/depromeet/domains/review/repository/ReviewRepository.java
@@ -1,0 +1,7 @@
+package com.depromeet.domains.review.repository;
+
+import com.depromeet.domains.review.entity.Review;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+}
