+++ server-domain/src/main/java/com/depromeet/domains/store/entity/Store.java
@@ -1,0 +1,47 @@
+package com.depromeet.domains.store.entity;
+
+import com.depromeet.domains.common.entity.BaseTimeEntity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Table(name = "Store")
+public class Store extends BaseTimeEntity {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long storeId;
+
+	@Column(nullable = false)
+	private Long categoryId;
+
+	@Column(nullable = false)
+	private String storeName;
+
+	@Column(nullable = false)
+	private Double latitude;
+
+	@Column(nullable = false)
+	private Double longitude;
+
+	private String thumbnailUrl;
+
+	private String jibunAddress;
+
+	private String roadAddress;
+
+	private String addressDetail;
+}
