+++ server-domain/src/main/java/com/depromeet/domains/store/repository/StoreQueryRepository.java
@@ -1,0 +1,18 @@
+package com.depromeet.domains.store.repository;
+
+import org.springframework.stereotype.Repository;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+import jakarta.persistence.EntityManager;
+
+@Repository
+public class StoreQueryRepository {
+	// queryDsl
+
+	private final JPAQueryFactory queryFactory;
+
+	public StoreQueryRepository(EntityManager entityManager) {
+		this.queryFactory = new JPAQueryFactory(entityManager);
+	}
+}
