+++ server-domain/src/main/java/com/depromeet/domains/store/repository/StoreRepository.java
@@ -1,0 +1,10 @@
+package com.depromeet.domains.store.repository;
+
+import com.depromeet.domains.store.entity.Store;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface StoreRepository extends JpaRepository<Store, Long> {
+
+}
