+++ server-domain/src/main/java/com/depromeet/domains/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.depromeet.domains.user.repository;
+
+import com.depromeet.domains.user.entity.User;
+import com.depromeet.enums.SocialType;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findBySocialTypeAndEmail(SocialType socialType, String email);
+
+}
