--- server-domain/src/main/java/com/depromeet/entity/EnumType.java
@@ -1,6 +1,0 @@
-package com.depromeet.entity;
-
-public interface EnumType {
-    String getName();
-    String getDescription();
-}
