--- server-domain/src/main/java/com/depromeet/entity/bookmark/Bookmark.java
@@ -1,28 +1,0 @@
-package com.depromeet.entity.bookmark;
-
-import com.depromeet.common.BaseTimeEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class Bookmark extends BaseTimeEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long bookmarkId;
-
-	@Column(nullable = false)
-	private String userId;
-
-	@Column(nullable = false)
-	private String storeId;
-}
