--- server-domain/src/main/java/com/depromeet/entity/db/config/JpaAuditingConfig.java
@@ -1,17 +1,0 @@
-package com.depromeet.entity.db.config;
-
-import java.util.Optional;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@Configuration
-public class JpaAuditingConfig implements AuditorAware<String> {
-
-	@Override
-	public Optional<String> getCurrentAuditor() {
-		return Optional.of("admin");
-	}
-}
