--- server-domain/src/main/java/com/depromeet/entity/follow/Follow.java
@@ -1,28 +1,0 @@
-package com.depromeet.entity.follow;
-
-import com.depromeet.common.BaseTimeEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class Follow extends BaseTimeEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long followId;
-
-	@Column(nullable = false)
-	private Long followerId;
-
-	@Column(nullable = false)
-	private Long followingId;
-}
