--- server-domain/src/main/java/com/depromeet/entity/store/StoreAppender.java
@@ -1,23 +1,0 @@
-package com.depromeet.entity.store;
-
-import org.springframework.stereotype.Component;
-
-import com.depromeet.entity.store.entity.Store;
-import com.depromeet.entity.store.persist.StorePersist;
-
-import lombok.RequiredArgsConstructor;
-
-@Component
-@RequiredArgsConstructor
-public class StoreAppender {
-	private final StorePersist storePersist;
-
-	public Long append(String storeName) {
-		Store store = Store
-			.builder()
-			.storeName(storeName)
-			.build();
-		return storePersist.append(store);
-	}
-
-}
