--- server-domain/src/main/java/com/depromeet/entity/store/entity/StoreRepository.java
@@ -1,9 +1,0 @@
-package com.depromeet.entity.store.entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface StoreRepository extends JpaRepository<Store, Long> {
-
-}
