+++ server-domain/src/main/java/com/depromeet/enums/EnumType.java
@@ -1,0 +1,6 @@
+package com.depromeet.enums;
+
+public interface EnumType {
+    String getName();
+    String getDescription();
+}
