+++ server-domain/src/main/java/com/depromeet/enums/Role.java
@@ -1,0 +1,13 @@
+package com.depromeet.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+
+	USER("ROLE_USER"), ADMIN("ROLE_ADMIN");;
+
+	private final String key;
+}
