+++ server-domain/src/main/java/com/depromeet/enums/SocialType.java
@@ -1,0 +1,5 @@
+package com.depromeet.enums;
+
+public enum SocialType {
+	APPLE, KAKAO
+}
