+++ server-domain/src/main/java/com/depromeet/test/Person.java
@@ -1,0 +1,20 @@
+package com.depromeet.test;
+
+public class Person {
+
+    private String name;
+    private String message;
+
+    public Person(String name, String message) {
+        this.name = name;
+        this.message = message;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
