--- server-domain/src/main/java/com/depromeet/user/enums/Role.java
@@ -1,13 +1,0 @@
-package com.depromeet.user.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-
-	USER("ROLE_USER"), ADMIN("ROLE_ADMIN");;
-
-	private final String key;
-}
