--- server-domain/src/main/java/com/depromeet/user/enums/SocialType.java
@@ -1,5 +1,0 @@
-package com.depromeet.user.enums;
-
-public enum SocialType {
-	APPLE, KAKAO
-}
